requirements.txt
altgraph==0.17.4
bottle==0.13.2
bottle-websocket==0.2.9
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
decorator==4.4.2
ffmpeg==1.4
filelock==3.16.1
fsspec==2024.9.0
future==1.0.0
gevent==24.10.3
gevent-websocket==0.10.1
greenlet==3.1.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==24.1
pefile==2023.2.7
pillow==11.0.0
proglog==0.1.10
PyAudio==0.2.14
pycparser==2.22
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pyparsing==3.2.0
pywin32-ctypes==0.2.3
regex==2024.9.11
requests==2.32.3
srt==3.5.3
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
vosk==0.3.45
websockets==13.1
whichcraft==0.6.1
zope.event==5.0
zope.interface==7.1.0
